pyproject.toml
[tool.poetry]
name = "sensormap-bff"
version = "0.1.0"
description = "The backend-for-frontend for sensormap"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
httpx = "^0.25.1"
python-keycloak = "^3.7.0"
fastapi-keycloak = "^1.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
